refactor(ui/screens): extract useApps, useSettings, useCurrentTheme hooks
Move data logic out of AppsScreen, SettingsScreen, and ThemeSelector
into dedicated hooks. Components become mostly JSX with no direct API
calls or state management inline.
